+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20200617135734.php
@@ -1,0 +1,26 @@
+<?php declare(strict_types=1);
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Add fields to block login after a certain number of failed attempts
+ */
+final class Version20200617135734 extends AbstractMigration
+{
+    public function up(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE admin ADD failed_authentication INT DEFAULT 0 NOT NULL, ADD last_failed_authentication DATETIME DEFAULT NULL');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE admin DROP failed_authentication, DROP last_failed_authentication');
+    }
+}